+++ src/components/triggers/operations/plugins/PluginNow.php
@@ -1,7 +1,10 @@
 <?php
 namespace deflou\components\triggers\operations\plugins;
 
+use deflou\interfaces\instances\IInstance;
 use deflou\interfaces\resolvers\events\IResolvedEvent;
+use deflou\interfaces\triggers\ITrigger;
+use deflou\interfaces\triggers\operations\ITriggerOperationPlugin;
 use deflou\interfaces\triggers\operations\plugins\IPluginDispatcher;
 
 class PluginNow implements IPluginDispatcher
@@ -9,5 +12,15 @@ class PluginNow implements IPluginDispatcher
     public function __invoke(string|int $triggerValue, IResolvedEvent $event): string|int
     {
         return preg_replace_callback('/\@now\((.*)\)\@/i', function ($matches) { return date($matches[1]);}, $triggerValue);
+    }
+
+    public function getTemplateData(IInstance $eventInstance, ITrigger $trigger, ITriggerOperationPlugin $plugin): array
+    {
+        return [
+            'Y.m.d H:i:s',
+            'd.m.Y H:i:s',
+            'Y.m.d',
+            'd.m.Y'
+        ];
     }
 }