+++ src/interfaces/triggers/events/ITriggerEventValuePlugin.php
@@ -1,6 +1,7 @@
 <?php
 namespace deflou\interfaces\triggers\events;
 
+use deflou\interfaces\applications\IHaveApplicationName;
 use extas\interfaces\IHasClass;
 use extas\interfaces\IHasDescription;
 use extas\interfaces\IHasName;
@@ -8,15 +9,11 @@ use extas\interfaces\IHaveUUID;
 use extas\interfaces\IItem;
 use extas\interfaces\parameters\IHaveParams;
 
-interface ITriggerEventValuePlugin extends IItem, IHaveUUID, IHasName, IHasDescription, IHasClass, IHaveParams
+interface ITriggerEventValuePlugin extends IItem, IHaveUUID, IHasName, IHasDescription, IHasClass, IHaveParams, IHaveApplicationName
 {
     public const SUBJECT = 'deflou.trigger.event.value.plugin';
 
     public const FIELD__APPLY_TO = 'apply_to';
-    public const FIELD__APPLICATION_NAME = 'app_name';
-
-    public function getApplicationName(): string;
-    public function setApplicationName(string $name): static;
 
     public function getApplyTo(): array;
     public function setApplyTo(array $applyTo): static;